refactor: muda o estilo da estilizacao para o styled components
